Added application load balancer pointing to Lavalink containers Added 2 lavalink servers running in AWS
Application load balancer acts as a single access point for both.
Without an ALB the IP will change every time the Lavalink container restarts.
If the IP changes, we need to change the IP address in the Bot and then deploy
for it to work again. The ALB acts as the access point every time the
containers start regardless of what IP they have. ALB routes all incomming
traffic to the Lavalink containers.